functional rest to kafka

+++ motion-service/motion-service-write/src/main/java/de/varsamis/motionservice/event/Projection3dEvent.java
@@ -3,13 +3,15 @@ package de.varsamis.motionservice.event;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class Projection3dEvent {
     private String device;
     private long timestamp;
-    private double x;
-    private double y;
-    private double z;
-    private double dist;
+    private BigDecimal x;
+    private BigDecimal y;
+    private BigDecimal z;
+    private BigDecimal dist;
 }
